delete more commands
