+++ editor/src/components/canvas/commands/utils/property-utils.ts
@@ -1,7 +1,7 @@
-import type { ElementPath, JSXElement } from 'utopia-shared/src/types'
+import type { ElementPath, JSXElement, PropertyPath } from 'utopia-shared/src/types'
 import { foldEither } from '../../../../core/shared/either'
 import type { ValueAtPath } from '../../../../core/shared/jsx-attributes'
-import { setJSXValuesAtPaths } from '../../../../core/shared/jsx-attributes'
+import { setJSXValuesAtPaths, unsetJSXValuesAtPaths } from '../../../../core/shared/jsx-attributes'
 import type { EditorState, EditorStatePatch } from '../../../editor/store/editor-state'
 import { modifyUnderlyingElementForOpenFile } from '../../../editor/store/editor-state'
 import { patchParseSuccessAtElementPath } from '../patch-utils'
@@ -41,6 +41,46 @@ export function applyValuesAtPath(
     }
   })
 
+  return {
+    editorStateWithChanges: workingEditorState,
+    editorStatePatch: editorStatePatch,
+  }
+}
+
+export function deleteValuesAtPath(
+  editorState: EditorState,
+  target: ElementPath,
+  properties: Array<PropertyPath>,
+): { editorStateWithChanges: EditorState; editorStatePatch: EditorStatePatch } {
+  const workingEditorState = modifyUnderlyingElementForOpenFile(
+    target,
+    editorState,
+    (element: JSXElement) => {
+      return foldEither(
+        () => {
+          return element
+        },
+        (updatedProps) => {
+          return {
+            ...element,
+            props: updatedProps,
+          }
+        },
+        unsetJSXValuesAtPaths(element.props, properties),
+      )
+    },
+  )
+
+  const editorStatePatch = patchParseSuccessAtElementPath(target, workingEditorState, (success) => {
+    return {
+      topLevelElements: {
+        $set: success.topLevelElements,
+      },
+      imports: {
+        $set: success.imports,
+      },
+    }
+  })
   return {
     editorStateWithChanges: workingEditorState,
     editorStatePatch: editorStatePatch,
