+++ editor/src/components/editor/actions/action-utils.ts
@@ -371,7 +371,6 @@ export function getElementsToNormalizeFromCommands(commands: CanvasCommand[]): E
       case 'ADJUST_CSS_LENGTH_PROPERTY':
       case 'SET_CSS_LENGTH_PROPERTY':
       case 'CONVERT_CSS_PERCENT_TO_PX':
-      case 'CONVERT_TO_ABSOLUTE':
         return command.target
       case 'ADD_CONTAIN_LAYOUT_IF_NEEDED':
       case 'SET_PROPERTY':
